Feature: Allow CORS for servers running on any port on localhost
